[FIX] Deploy on university server
Processor on our target server is very, very old. It's actually impossible to build latest version of torch on it.
Building docker image locally is not a solution - we hit missing functions in runtime. This causes hella lot of core dumps.
Fortunatelly, openface has dockerized version build with some bizzare flags that actually can be run on uni.